New TestVersionTrait
Move the logic related to retrieving the `testVersion` config variable and whether certain PHP versions should be supported to a dedicated trait.

This will allow for sniffs which don't extend the base PHPCompatibility `Sniff` class, like sniffs which will be switched over to using abstract base sniff classes from PHPCSUtils, to use these methods as well.

Includes splitting off the unit tests for these methods to their own dedicated test file.

Includes updating the PHPUnit config to check for code coverage of Traits.
